# Conflicts:
#	Core/WebAPI/Clients/CatalogClient.cs
#	Core/WebAPI/Controllers/CollectionController.cs
#	Core/WebAPI/Controllers/PersonalListController.cs
#	Core/WebAPI/Models/API/KalturaTvmDeviceRule.cs
#	Core/WebAPI/Models/API/KalturaTvmGeoRule.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByAudioCodecAction.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByFileTypeIdAction.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByFileTypeIdForAssetTypeAction.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByLabelAction.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByQualityAction.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByStreamerTypeAction.cs
#	Core/WebAPI/Models/ConditionalAccess/FilterActions/Files/KalturaFilterFileByVideoCodecAction.cs
#	Core/WebAPI/Models/Pricing/KalturaCollection.cs

+++ Core/WebAPI/Models/IngestStatus/KalturaIngestStatusEpgConfiguration.cs
@@ -1,8 +1,5 @@
 ﻿using Newtonsoft.Json;
-using System;
-using System.Collections.Generic;
 using System.Runtime.Serialization;
-using System.Text;
 using System.Xml.Serialization;
 using WebAPI.Managers.Scheme;
 using WebAPI.Models.General;
@@ -27,6 +24,6 @@ namespace WebAPI.Models.IngestStatus
         [JsonProperty("retainingPeriod")]
         [XmlElement(ElementName = "retainingPeriod")]
         [SchemeProperty(IsNullable = true, MinInteger = 0)]
-        public int? RetainingPeriod { get; set; }
+        public long? RetainingPeriod { get; set; }
     }
 }
